turn WeakObserver into a regular class instead of an interface
