+++ src/main/java/sprouts/impl/PropertyChangeListeners.java
@@ -36,8 +36,8 @@ public final class PropertyChangeListeners<T>
     }
 
     public void onChange( Observer observer ) {
-        if ( observer instanceof WeakObserverImpl)
-            this.onChange(Sprouts.factory().defaultObservableChannel(), new WeakObserverAsActionImpl<>( (WeakObserverImpl<?>) observer ) );
+        if ( observer instanceof WeakObserver)
+            this.onChange(Sprouts.factory().defaultObservableChannel(), new WeakObserverAsActionImpl<>( (WeakObserver<?>) observer ) );
         else
             this.onChange(Sprouts.factory().defaultObservableChannel(), new ObserverAsActionImpl<>(observer) );
     }
