+++ src/main/java/sprouts/impl/SproutsFactory.java
@@ -223,8 +223,6 @@ public interface SproutsFactory
 
     <O,D> WeakAction<O,D> actionOfWeak(O owner, BiConsumer<O, D> action );
 
-    <O> WeakObserver<O> observerOfWeak( O owner, Consumer<O> action );
-
     /**
      *   The default id for properties which do not have an id explicitly specified.
      *   The id of a property is used to identify it in the system or as part of a view model
