+++ src/test/groovy/sprouts/Property_Binding_Spec.groovy
@@ -4,6 +4,7 @@ import spock.lang.Narrative
 import spock.lang.Specification
 import spock.lang.Subject
 import spock.lang.Title
+import sprouts.impl.WeakObserver
 
 import java.lang.ref.WeakReference
 
@@ -260,7 +261,7 @@ class Property_Binding_Spec extends Specification
         and : 'A trace list to record the side effects.'
             var trace = []
         and : 'Finally we register a weak observer on the property.'
-            Viewable.cast(property).subscribe(Observer.ofWeak(owner,{trace << "!"}))
+            Viewable.cast(property).subscribe(new WeakObserver(owner,{trace << "!"}))
 
         when : 'We change the property.'
             property.set(43)
